pyproject.toml
[tool.poetry]
name = "tthb"
version = "0.1.0"
description = ""
authors = ["Amy Langley <amy.a.langley@gmail.com>"]
repository = "https://github.com/amy-langley/tracking-trans-hate-bills"
packages = [
    { include = "lib", from = "." },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
jupyterlab = "^3.6.1"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
jsonpath = "^0.82"
plotly-express = "^0.4.1"
wordcloud = "^1.8.2.2"
nltk = "^3.8.1"
mergedeep = "^1.3.4"
pyjsparser = "^2.7.1"
requests-cache = "^1.0.0"
pdfplumber = "^0.8.0"
pillow = "^9.4.0"
kaleido = "0.2.0"
snakemake = "^7.24.2"
typer = "^0.7.0"
types-beautifulsoup4 = "^4.12.0.0"
pandas-stubs = "^1.5.3.230321"
types-requests = "^2.28.11.15"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"

[tool.mypy]
[[tool.mypy.overrides]]
module = [
    "nltk.*",
    "pyjsparser.*",
    "mergedeep.*",
]
ignore_missing_imports = true

[tool.pylint]
[tool.pylint.master]
disable=["C0114", "W1203"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"